Move network calls into static functions
