+++ Basic-Car-Maintenance/Shared/Localizable.xcstrings
@@ -2827,9 +2827,6 @@
         }
       }
     },
-    "Maintenance Vehicle" : {
-
-    },
     "Make" : {
       "localizations" : {
         "be" : {
